Fixes #13 where portal state not updated when wifi connects but was already enabled
When PortalDetectorService called isWifiConnectedFromContext(), the
ConnectivityManager was reporting the networking type as not wifi (cell
network or something). This seems to be a bug in the framework.

The workaround is to grab the wifi state from the broadcast intent
initiating the state update, and pass it (via the service start intent)
to the service. The service then checks this first, and if the start
intent suggests that wifi is already connected, it doesn't perform its
own check. If no suggestion was given, the service checks using its
context.